pyproject.toml
[project]
name = "MaiBotDiscordAdapter"
version = "1.0.0"
description = "A MaiBot adapter for Discord"

license = {text = "GPL-v3.0"}
readme = "README.md"
dependencies = [
    "discord.py[voice]>=2.3.0",
    "discord-ext-voice-recv>=0.4.1a139",
    "maim_message",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
    "aiofiles",
    "tomlkit",
    "asyncio",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
]

[tool.ruff]

include = ["*.py"]

# 行长度设置
line-length = 120

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []

# 启用的规则
select = [
    "E", # pycodestyle 错误
    "F", # pyflakes
    "B", # flake8-bugbear
]

ignore = ["E711","E501"]

[tool.ruff.format]
docstring-code-format = true
indent-style = "space"

# 使用双引号表示字符串
quote-style = "double"

# 尊重魔法尾随逗号
skip-magic-trailing-comma = false

# 自动检测合适的换行符
line-ending = "auto"